ai-services/RAG-vid/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# YouTube processing
youtube-transcript-api==0.6.3
yt-dlp==2024.12.13

# Video & Audio Processing
ffmpeg-python==0.2.0
pydub==0.25.1
librosa==0.10.2
moviepy==1.0.3

# Speech Recognition
openai-whisper==20231117
SpeechRecognition==3.10.0
pyaudio==0.2.13

# Machine Learning
sentence-transformers==2.6.1        # ✅ đã fix lỗi cached_download
transformers==4.36.0
huggingface-hub==0.20.3   # 👈 Hạ xuống phiên bản tương thích
torch==2.1.0
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.0

# Gemini
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
pathlib==1.0.1
asyncio==3.4.3
aiofiles==23.2.1
